python_code(production)/requirements.txt
Flask==3.0.0
openai==1.12.0
gspread==6.0.0
google-auth==2.27.0
python-dotenv==1.0.0
requests==2.31.0
